Restructure of public key -> peer state
Restructured the mapping of public keys to peer state in the project.
The handshake device is now generic over an opaque type,
which enables it to be the sole place where public keys are mapped to
the peer states.

This gets rid of the "peer" map in the WireGuard devices
and avoids having to include the public key in the handshake peer state.